refactor out shape tools from view-controller,
